삭제 된 파일 복구 및 admin폴더 이름 변경입니다.
