+++ src/main/reactfront/src/pages/Signup/SignupPage.tsx
@@ -13,11 +13,11 @@ import KeyIcon from '@mui/icons-material/Key';
 import PersonIcon from '@mui/icons-material/Person';
 import PhoneIphoneIcon from '@mui/icons-material/PhoneIphone';
 import CalendarTodayIcon from '@mui/icons-material/CalendarToday';
+import GenericButton from '../../components/common/genericButton/GenericButton';
 import {
   SignupButtons,
   SignupForm,
-} from '../../styles/Signup/SignupPage.styles';
-import GenericButton from '../../components/common/genericButton/GenericButton';
+} from '../../styles/signup/SignupPage.styles';
 const SignupPage: React.FC = () => {
   const navigate = useNavigate();
