--- src/main/reactfront/src/styles/Signup/SignupPage.ts
@@ -1,19 +1,0 @@
-import styled from 'styled-components';
-
-export const SignupForm = styled.div`
-  width: 500px;
-  height: 100vh;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center; 
-`;
-
-export const SignupButtons = styled.div`
-  margin-top: 14px;
-  display: flex;
-  justify-content: space-between;
-  height: 45px;
-  width: 100%;
-`;
