+++ src/main/reactfront/src/styles/administrator/AdminPage.styles.ts
