+++ src/main/reactfront/src/styles/administrator/CommunityListPage.styles.ts
