+++ src/main/reactfront/src/styles/administrator/EtcInquiryListPage.styles.ts
