+++ src/main/reactfront/src/styles/administrator/ItemInquiryListPage.styles.ts
