+++ src/main/reactfront/src/styles/administrator/ItemListPage.styles.ts
