+++ src/main/reactfront/src/styles/administrator/PostListPage.styles.ts
