+++ src/main/reactfront/src/types/administrator/CommunityData.tsx
