+++ src/main/reactfront/src/types/administrator/EtcInquiryDatat.tsx
