+++ src/main/reactfront/src/types/administrator/InformationData.tsx
