+++ src/main/reactfront/src/types/administrator/ItemInquiryData.tsx
